action button is now a component
